% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{as_utf8_text_bold}
\alias{as_utf8_text_bold}
\title{Format text bold (sans) using unicode}
\usage{
as_utf8_text_bold(text)
}
\arguments{
\item{text}{A text (as string)}
}
\value{
Unicode text
}
\description{
Format text bold (sans) using unicode
}
\examples{
as_utf8_text_bold("Hello World")
}
