% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{as_utf8_text_circled}
\alias{as_utf8_text_circled}
\title{Format text circled using unicode}
\usage{
as_utf8_text_circled(text)
}
\arguments{
\item{text}{A text (as string)}
}
\value{
Unicode text
}
\description{
Format text circled using unicode
}
\examples{
as_utf8_text_circled("Hello World")
}
