% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{as_utf8_text_cursive}
\alias{as_utf8_text_cursive}
\title{Format text cursive using unicode}
\usage{
as_utf8_text_cursive(text)
}
\arguments{
\item{text}{A text (as string)}
}
\value{
Unicode text
}
\description{
Format text cursive using unicode
}
\examples{
as_utf8_text_cursive("Hello World")
}
