% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{as_utf8_text_gothic}
\alias{as_utf8_text_gothic}
\title{Format text gothic using unicode}
\usage{
as_utf8_text_gothic(text)
}
\arguments{
\item{text}{A text (as string)}
}
\value{
Unicode text
}
\description{
Format text gothic using unicode
}
\examples{
as_utf8_text_gothic("Hello World")
}
