% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{as_utf8_text_italic}
\alias{as_utf8_text_italic}
\title{Format text italic using unicode}
\usage{
as_utf8_text_italic(text)
}
\arguments{
\item{text}{A text (as string)}
}
\value{
Unicode text
}
\description{
Format text italic using unicode
}
\examples{
as_utf8_text_italic("Hello World")
}
