% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_block}
\alias{utf8_block}
\title{Block from the UTF8 table}
\usage{
utf8_block(from = "0x1F600", to = "0x1F64F", show_names = FALSE)
}
\arguments{
\item{from}{Hexcode starting}

\item{to}{Hexcode end}
}
\value{
Vector of UTF8 characters
}
\description{
Block from the UTF8 table
}
\examples{
utf8_block(from ="0x1F600", to="0x1F64F")
}
