% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_collection}
\alias{utf8_collection}
\title{Collection of unicode symbols}
\usage{
utf8_collection()
}
\arguments{
\item{text}{A text (as string)}
}
\value{
Unicode text
}
\description{
Collection of unicode symbols
}
\examples{
utf8_collection()
}
