% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_inspect}
\alias{utf8_inspect}
\title{Inspect a Utf8 character}
\usage{
utf8_inspect(chr)
}
\arguments{
\item{chr}{A Utf8 character}
}
\value{
Prints infos int & hex
}
\description{
Inspect a Utf8 character
}
\examples{
u <- utf8_collection()
utf8_inspect(u$fav$heart)
}
