% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_number_circle}
\alias{utf8_number_circle}
\title{Format number circled using unicode}
\usage{
utf8_number_circle(num)
}
\arguments{
\item{num}{Number}
}
\value{
Unicode text
}
\description{
Format number circled using unicode
}
\examples{
utf8_number_box(123.45)
}
