% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_number_heart}
\alias{utf8_number_heart}
\title{Format number as hearts using unicode}
\usage{
utf8_number_heart(num = 5, max_num = 10)
}
\arguments{
\item{num}{Number}

\item{max_num}{Maximum number}
}
\value{
Unicode text
}
\description{
Format number as hearts using unicode
}
\examples{
utf8_number_heart(3)
}
