% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_number_star}
\alias{utf8_number_star}
\title{Format number as stars using unicode}
\usage{
utf8_number_star(num = 5, max_num = 10)
}
\arguments{
\item{num}{Number}

\item{max_num}{Maximum number}
}
\value{
Unicode text
}
\description{
Format number as stars using unicode
}
\examples{
utf8_number_box(123.45)
}
