% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_text_bolditalic}
\alias{utf8_text_bolditalic}
\title{Format text bold and italic using unicode}
\usage{
utf8_text_bolditalic(text)
}
\arguments{
\item{text}{A text (as string)}
}
\value{
Unicode text
}
\description{
Format text bold and italic using unicode
}
\examples{
utf8_text_bolditalic("Hello World")
}
