% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_text_circle}
\alias{utf8_text_circle}
\title{Format text circled using unicode}
\usage{
utf8_text_circle(text)
}
\arguments{
\item{text}{A text (as string)}
}
\value{
Unicode text
}
\description{
Format text circled using unicode
}
\examples{
utf8_text_circle("Hello World")
}
