% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_text_italic}
\alias{utf8_text_italic}
\title{Format text italic using unicode}
\usage{
utf8_text_italic(text)
}
\arguments{
\item{text}{A text (as string)}
}
\value{
Unicode text
}
\description{
Format text italic using unicode
}
\examples{
utf8_text_italic("Hello World")
}
