% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_text_monospace}
\alias{utf8_text_monospace}
\title{Format text monospace using unicode}
\usage{
utf8_text_monospace(text)
}
\arguments{
\item{text}{A text (as string)}
}
\value{
Unicode text
}
\description{
Format text monospace using unicode
}
\examples{
utf8_text_monospace("Hello World")
}
