% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ify.R
\name{utf8_collection}
\alias{utf8_collection}
\title{Collection of unicode symbols}
\usage{
utf8_collection()
}
\value{
List of unicode collections
}
\description{
Collection of unicode symbols
}
\examples{
# show complete collection (as list)
utf8_collection()

# show fav collection 
u <- utf8_collection()
unlist(u$fav)

# show single character
u$fav$heart
}
